Update dependencies: colored 3.1, indicatif 0.18, scraper 0.27, rusqlite 0.39, tokio-tungstenite 0.29, x509-parser 0.18

+++ db/store.rs
@@ -53,7 +53,7 @@ pub fn save_scan(conn: &Connection, result: &ScanResult) -> Result<(), rusqlite:
             now,
             result.scan_version,
             result.intensity_level,
-            result.statistics.total,
+            result.statistics.total as i64,
             result.duration_secs,
         ],
     )?;
